src/main/java/items/SheetMusic.java
package items;

public class SheetMusic {

    private String description;
    private double buyingPrice;
    private double sellingPrice;

    public SheetMusic(String description, double buyingPrice, double sellingPrice) {
        this.description = description;
        this.buyingPrice = buyingPrice;
        this.sellingPrice = sellingPrice;
    }

    public String getDescription() {
        return description;
    }

    public double getBuyingPrice() {
        return buyingPrice;
    }

    public double getSellingPrice() {
        return sellingPrice;
    }
}
